requirements.txt
Flask==2.3.2
flask-socketio==5.3.3
gunicorn==20.2.0
gevent==22.10.2
discord.py==2.0.0
python-dotenv==1.0.0
